+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ANSI escape sequences to colorize output and the `--no-colors` CLI option to disable it
 - Driver in a single-contract mode tries to copy all the .tact and .fc files to resolve imports
 - Short CLI options. See: https://nowarp.io/tools/misti/docs/next/tutorial/cli
+- CLI: `--min-severity/-m` option
 
 ### Changed
 - Warnings now have more comprehensive descriptions and are sorted by severity